+++ robot/src/components/Library.js
@@ -1,6 +1,6 @@
 import React from "react";
 import LibrarySong from "./LibrarySong";
-const Library = ({ songs, setCurrentSong }) => {
+const Library = ({ songs, setCurrentSong  , audioRef }) => {
   return (
     <div className="library">
       <h2>Library</h2>
@@ -10,6 +10,8 @@ const Library = ({ songs, setCurrentSong }) => {
             song={song}
             setCurrentSong={setCurrentSong}
             key={song.id}
+            audioRef={audioRef}
+           
           />
         ))}
       </div>
