+++ robot/src/container/App.js
@@ -1,4 +1,4 @@
-import React, { useState, useMemo } from "react";
+import React, { useState, useRef } from "react";
 import Player from "../components/Player";
 import Song from "../components/Song";
 import "../styles/app.scss";
@@ -9,6 +9,27 @@ const App = () => {
   const [songs, setSongs] = useState(data());
   const [currentSong, setCurrentSong] = useState(songs[0]);
   const [isPlaying, setIsPlaying] = useState(false);
+  // Ref
+  const audioRef = useRef(null);
+
+  //time update handler
+
+  const timeUpdateHandler = (e) => {
+    const current = e.target.currentTime;
+    const duration = e.target.duration;
+
+    setSongInfo({
+      ...songInfo,
+      currentTime: current,
+      duration,
+    });
+  };
+  //state
+
+  const [songInfo, setSongInfo] = useState({
+    currentTime: 0,
+    duration: 0,
+  });
 
   return (
     <div className="App">
@@ -17,8 +38,16 @@ const App = () => {
         currentSong={currentSong}
         isPlaying={isPlaying}
         setIsPlaying={setIsPlaying}
+        audioRef={audioRef}
+        setSongInfo={setSongInfo}
+        songInfo={songInfo}
+        timeUpdateHandler={timeUpdateHandler}
       />
-      <Library songs={songs} setCurrentSong={setCurrentSong}/>
+      <Library
+        songs={songs}
+        setCurrentSong={setCurrentSong}
+        isPlaying={isPlaying}
+      />
     </div>
   );
 };
